update: move to utils function and clean up
