+++ packages/sdk/src/transactionManager.ts
@@ -4,29 +4,8 @@ import { PrepareParams, CancelParams, FulfillParams, getUuid, isNode } from "@co
 import { TransactionManager as TTransactionManager, IERC20Minimal } from "@connext/nxtp-contracts/typechain";
 import TransactionManagerArtifact from "@connext/nxtp-contracts/artifacts/contracts/TransactionManager.sol/TransactionManager.json";
 import ERC20 from "@connext/nxtp-contracts/artifacts/contracts/interfaces/IERC20Minimal.sol/IERC20Minimal.json";
-import _contractDeployments from "@connext/nxtp-contracts/deployments.json";
 
 import { ChainNotConfigured } from "./error";
-
-export const contractDeployments = _contractDeployments;
-
-/**
- * Returns the address of the `TransactionManager` deployed to the provided chain, or undefined if it has not been deployed
- *
- * @param chainId - The chain you want the address on
- * @returns The deployed address or `undefined` if it has not been deployed yet
- */
-export const getDeployedTransactionManagerContractAddress = (chainId: number): string | undefined => {
-  const record = (contractDeployments as any)[String(chainId)] ?? {};
-  const name = Object.keys(record)[0];
-  if (!name) {
-    return undefined;
-  }
-
-  const address = record[name]?.contracts?.TransactionManager?.address;
-
-  return address;
-};
 
 /**
  * @classdesc Multi-chain wrapper around TranasctionManager contract interactions
