+++ packages/utils/src/chainData.ts
@@ -1,3 +1,5 @@
+import _contractDeployments from "@connext/nxtp-contracts/deployments.json";
+
 export type ChainData = {
   name: string;
   chainId: number;
@@ -27,4 +29,28 @@ export type ChainData = {
     icon: string;
     standard: string;
   }[];
+};
+
+export const contractDeployments = _contractDeployments;
+
+/**
+ * Returns the address of the `TransactionManager` deployed to the provided chain, or undefined if it has not been deployed
+ *
+ * @param chainId - The chain you want the address on
+ * @returns The deployed address or `undefined` if it has not been deployed yet
+ */
+export const getDeployedTransactionManagerContract = (chainId: number): { address: string; abi: any } | undefined => {
+  const record = (contractDeployments as any)[String(chainId)] ?? {};
+  const name = Object.keys(record)[0];
+  if (!name) {
+    return undefined;
+  }
+
+  const contract = record[name]?.contracts?.TransactionManager;
+
+  if (contract) {
+    return { address: contract.address, abi: contract.abi };
+  }
+
+  return undefined;
 };
